+++ SubscriptionForm/src/components/Terms.js
@@ -10,11 +10,11 @@ const Terms = () => {
   };
 
   return (
-    <div className="col-12-xs pt-2 pb-2">
+    <div className="TOS">
       <Checkbox value={checked} onChange={handleChange} />
-      <p className="ml-4 text-gray-1">
-        I agree to{" "}
-        <span className="btn-black underline ">terms of service</span>
+      <p className="i-agree-to-terms-of">
+        <span className="text-wrapper">I agree to </span>
+        <span className="span">terms of service</span>
       </p>
     </div>
   );