+++ Changelog.md
@@ -1,5 +1,6 @@
 # ToDoList
 1. Implement Anima - Figma to React implementation to simplify SASS implementation for responsive styling
+2. Update submit buttons SVG
 
 # Done list
 Simple log - place cursor over the change you are interested in to get information about contributor and commits time (works in VS Code if GitLens extension is used)